setup action to modify foreign entry permission
