+++ client/src/redux/actions/profile.js
@@ -24,30 +24,6 @@ export const getMyProfileStuffAction = () => {
     }
 }
 
-export const getForeignEntriesAction = () => {
-	return async (dispatch, getState) => {
-        const userId = getState().profile.user.id
-        const token = getState().profile.token
-
-        const headers = {
-            'Content-Type': 'application/json',
-            'Authorization': `bearer ${token}`
-        }
-
-        try {
-            const response = await axios.get(`http://localhost:3001/api/profile/${userId}/manage`, {headers: headers})
-
-            dispatch({
-                type: 'FOREIGN_ENTRIES',
-                payload: response.data
-            })
-        }
-        catch(error){
-            console.log(error)
-        }
-    }
-}
-
 export const storedProfileAction = formInfo => {
 	return async (dispatch, getState) => {
         const retrievedProfile = localStorage.getItem('profile');
